+++ packages/cms/cmsfilter/src/index.ts
@@ -1,10 +1,12 @@
-import { assessScript } from '$utils/attributes';
+import { assessScript, initAttributes } from '$utils/attributes';
 import { importAnimations, importCMSCore } from '$utils/import';
 import { init } from './init';
 
 /**
  * Init
  */
+initAttributes();
+
 const { currentScript } = document;
 const { preventsLoad } = assessScript(currentScript);