+++ utils/import.ts
@@ -4,8 +4,7 @@ import type { AnimationImport } from 'packages/animation/src/types';
 import type { CMSCoreImport } from 'packages/cms/cmscore/src/types';
 
 const ANIMATIONS_SOURCE = 'https://cdn.jsdelivr.net/npm/@finsweet/attributes-animation@1.0.1/functions.js';
-const CMS_CORE_SOURCE =
-  ' https://onedrive.live.com/download?cid=6BF5AC2A9E9521CC&resid=6BF5AC2A9E9521CC%21268836&authkey=AL8OuJL7uMsa2es';
+const CMS_CORE_SOURCE = 'https://cdn.jsdelivr.net/npm/@finsweet/attributes-cmscore@1.0.0/cmscore.js';
 
 /**
  * Dynamically imports the `animation` package.
@@ -36,6 +35,8 @@ export const importAnimations = async (): AnimationImport => {
  */
 export const importCMSCore = async (): CMSCoreImport => {
   const { fsAttributes } = window;
+
+  fsAttributes.cms ||= {};
 
   if (fsAttributes.cms.coreImport) return fsAttributes.cms.coreImport;