+++ packages/displayvalues/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `displayvalues` Changelog
 
+## [v1.0.2] 25th October 2021
+
+- Added `initAttributes` call to make sure `window.fsAttributes` exists.
+
 ## [v1.0.1] 20th October 2021
 
 - Made sure a source element has targets on initialization.