+++ packages/linkblockedit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `linkblockedit` Changelog
 
+## [v1.0.5] 25th October 2021
+
+- Added `initAttributes` call to make sure `window.fsAttributes` exists.
+
 ## [v1.0.4] 8th October 2021
 
 - Updated `examples.json`.