+++ packages/mirrorclick/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `mirrorclick` Changelog
 
+## [v1.1.3] 25th October 2021
+
+- Added `initAttributes` call to make sure `window.fsAttributes` exists.
+
 ## [v1.1.2] 8th October 2021
 
 - Updated `examples.json`.