+++ packages/populate/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `populate` Changelog
 
-## [v1.0.0] Date
+## [v1.0.1] 25th October 2021
+
+- Added `initAttributes` call to make sure `window.fsAttributes` exists.
+
+## [v1.0.0] 18th October 2021
 
 - Created the attribute package.