+++ packages/scrolldisable/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `scrolldisable` Changelog
 
+## [v1.1.9] 25th October 2021
+
+- Added `initAttributes` call to make sure `window.fsAttributes` exists.
+
 ## [v1.1.8] 8th October 2021
 
 - Updated `examples.json`.