+++ packages/untransform/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `untransform` Changelog
 
+## [v1.0.8] 25th October 2021
+
+- Added `initAttributes` call to make sure `window.fsAttributes` exists.
+
 ## [v1.0.7] 16th October 2021
 
 - Updated `examples.json`.